...G...PS. [ZBX-17036] fixed memory leak and wrong type cast

+++ src/libs/zbxcommon/file.c
@@ -106,7 +106,7 @@ int	zbx_read(int fd, char *buf, size_t count, const char *encoding)
 	if ((zbx_offset_t)-1 == (offset = zbx_lseek(fd, 0, SEEK_CUR)))
 		return -1;
 
-	if (0 >= (nbytes = read(fd, buf, (unsigned int)count)))
+	if (0 >= (ssize_t)(nbytes = read(fd, buf, (unsigned int)count)))
 		return (int)nbytes;
 
 	find_cr_lf_szbyte(encoding, &cr, &lf, &szbyte);
